rename: 스키마 한 파일에 작성
